Refactored code and configuration
